Server/requirements.txt
flask
Pillow
firebase-admin
